% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaigns.R
\name{dg_campaigns}
\alias{dg_campaigns}
\title{DG Campaigns}
\usage{
dg_campaigns()
}
\value{
a tibble of the org's dogooder campaigns
}
\description{
Returns information about all campaigns in an account. Information includes title, description, urls, actions, etc.
}
