% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_wrangling.R
\name{dg_unnest}
\alias{dg_unnest}
\title{Unnest actions/campaigns data}
\usage{
dg_unnest(x)
}
\arguments{
\item{x}{table with nested columns}
}
\value{
wider table
}
\description{
Used to turn columns containing tables (eg. campaign, action, form_data), each of which may have different dimensions, into wider tables with new columns using names with the format prior_colname.colname_of_nested_table eg. action turns into action.id and action.title
}
\examples{

}
