% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogoodr-package.R
\docType{package}
\name{dogoodr-package}
\alias{dogoodr}
\alias{dogoodr-package}
\title{dogoodr: Interface to the 'Do Gooder' API}
\description{
Interact directly with the 'Do Gooder' <https://dogooder.co/> API from R.
}
\author{
\strong{Maintainer}: Danny Smith \email{danny@gorcha.org}

Authors:
\itemize{
  \item Mark Brisbane \email{mark.brisbane@amnesty.org.au}
}

}
\keyword{internal}
