% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{first_moment}
\alias{first_moment}
\alias{this_morning}
\alias{days_ago}
\alias{yesterday}
\alias{weeks_ago}
\alias{ever}
\title{Convenience functions dg_actions "since" argument}
\usage{
first_moment(day)

this_morning()

days_ago(n)

yesterday()

weeks_ago(n)

ever()
}
\arguments{
\item{day}{a date value}

\item{n}{an integer for the number of days or weeks previous to this moment}
}
\value{
a datetime object
}
\description{
Convenience functions dg_actions "since" argument
}
\section{Functions}{
\itemize{
\item \code{this_morning}: returns first moment of the current day

\item \code{days_ago}: returns first moment n days ago

\item \code{yesterday}: returns first moment yesterday

\item \code{weeks_ago}: returns first moment n weeks ago

\item \code{ever}: returns first moment of the year 2000
}}

