% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{flat_values}
\alias{flat_values}
\title{Flat list values}
\usage{
flat_values(x)
}
\arguments{
\item{x}{a list}
}
\value{
list items without sublists
}
\description{
Returns list items within a level that aren't lists
}
