% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{parse_response.actionfeed}
\alias{parse_response.actionfeed}
\title{Parse response list from actionfeed}
\usage{
\method{parse_response}{actionfeed}(x)
}
\arguments{
\item{x}{response list}
}
\value{
tibble
}
\description{
Parse response list from actionfeed
}
