% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{parse_response}
\alias{parse_response}
\title{Parse a response list into a rectangle}
\usage{
parse_response(x)
}
\arguments{
\item{x}{a response list}
}
\value{
a tibble based off the out_class of a response list
}
\description{
Parse a response list into a rectangle
}
