% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polite_message}
\alias{polite_message}
\title{Concatenate and post to console}
\usage{
polite_message(msg, silent = TRUE, env = parent.frame())
}
\arguments{
\item{msg}{character with concatenations in curly brackets}

\item{silent}{whether to message}

\item{env}{environment with variables to concatenate}
}
\value{
a console log
}
\description{
Applies glue to the string it receives, and returns if silent=FALSE
}
