% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{process_pages}
\alias{process_pages}
\title{Process remotely paginated dogooder data}
\usage{
process_pages(parsed_response, max_requests = NULL, silent = FALSE, ...)
}
\arguments{
\item{parsed_response}{previous response to pull the 'next' value from}

\item{max_requests}{stops early if it exceeds this number}

\item{silent}{whether to print request details to the console}

\item{...}{other parameters for the next request}
}
\value{
parsed responses binded together
}
\description{
Process remotely paginated dogooder data
}
