% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{response_parse}
\alias{response_parse}
\title{Parse http response json to a list}
\usage{
response_parse(response, out_class)
}
\arguments{
\item{response}{http response in json format}

\item{out_class}{class to assign the response list}
}
\value{
response list
}
\description{
Parse http response json to a list
}
