% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{stack_lists}
\alias{stack_lists}
\title{Stacks several lists into tibble rows}
\usage{
stack_lists(l)
}
\arguments{
\item{l}{a list where each item is a list, which will become a row}
}
\value{
a tibble
}
\description{
Requires a list of lists as an input. Tries to turn sublists into lists as
well.
}
