% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{url_args}
\alias{url_args}
\title{Extract a URLs arguments for new requests}
\usage{
url_args(url)
}
\arguments{
\item{url}{character item with arguments embedded in it}
}
\value{
a list of arguments
}
\description{
Extract a URLs arguments for new requests
}
