% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{dg_actions}
\alias{dg_actions}
\title{DG Actions}
\usage{
dg_actions(
  since = this_morning(),
  campaign_id = NULL,
  starting_with = c("earliest", "latest"),
  process_pagination = FALSE
)
}
\arguments{
\item{since}{time object representing the earliest action to be returned}

\item{campaign_id}{vector of campaign id integers to be returned}

\item{starting_with}{should the return list be arranged to start with the earliest or latest action}

\item{process_pagination}{if TRUE function will make several requests to download each page of actions, or should it only return the first 100}
}
\value{
tibble with nested columns for campaign, action, form_data
}
\description{
Returns a feed the latest action log entries for the entire account. It includes the campaign, action, and form data submitted by the supporter.
}
\details{
Useful convenience functions for 'since' argument include ever() which returns the first second of the CE, this_morning() which returns the first second of today first_moment() which returns the first second of any day, yesterday(), days_ago(), weeks_ago() which behave similarly.
}
\examples{
minimum_time <- lubridate::now() - lubridate::days(5)
dg_actions(since=minimum_time, campaign_id=5618, starting_with="latest", process_pagination=FALSE)
}
