% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{dg_api}
\alias{dg_api}
\title{Make a DG API Request}
\usage{
dg_api(
  endpoint,
  out_class,
  clean_response = FALSE,
  query_param = list(),
  process_pagination = FALSE
)
}
\arguments{
\item{endpoint}{name of endpoint}

\item{out_class}{class to use for the parse_response request}

\item{clean_response}{whether to run parse_response on the request}

\item{query_param}{URL parameters to send as part of the request}

\item{process_pagination}{if there are multiple response pages, should
whether to send multiple requests}
}
\value{
a response list
}
\description{
Lower level function used by functions with dg_ prefix.
}
\examples{
\dontrun{
dg_api(
  endpoint = "action-log-feed",
  out_class = "actionfeed",
  query_param = list(campaign_ids = 5618),
  clean_response = TRUE,
  process_pagination = TRUE
)
}
}
