% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{dg_api}
\alias{dg_api}
\title{Make a DG API Request}
\usage{
dg_api(
  endpoint,
  out_class,
  clean_response = FALSE,
  query_param = list(),
  process_pagination = FALSE,
  ...
)
}
\arguments{
\item{endpoint}{name of endpoint}

\item{out_class}{class to use for the parse_response request}

\item{clean_response}{whether to run parse_response on the request, or}

\item{...}{}
}
\value{

}
\description{
Lower level function used by functions with dg_ prefix.
}
\examples{
dg_api(endpoint="activity_feed", out_class="", clean_response=TRUE, query_param=, process_pagination=FALSE)
}
